* commit 'a7b364cfd4aa9584124284d544f3ae9b11b494b4':
  Timestamp synchronization must use 64-bit long to avoid overflow.

+++ tests/tests/hardware/src/android/hardware/cts/helpers/sensorverification/EventTimestampSynchronizationVerification.java
@@ -67,10 +67,10 @@ public class EventTimestampSynchronizationVerification extends AbstractSensorVer
      */
     public static EventTimestampSynchronizationVerification getDefault(
             TestSensorEnvironment environment) {
-        int reportLatencyUs = environment.getMaxReportLatencyUs();
-        int fifoMaxEventCount = environment.getSensor().getFifoMaxEventCount();
+        long reportLatencyUs = environment.getMaxReportLatencyUs();
+        long fifoMaxEventCount = environment.getSensor().getFifoMaxEventCount();
         if (fifoMaxEventCount > 0) {
-            int fifoBasedReportLatencyUs =
+            long fifoBasedReportLatencyUs =
                     fifoMaxEventCount * environment.getMaximumExpectedSamplingPeriodUs();
             reportLatencyUs = Math.min(reportLatencyUs, fifoBasedReportLatencyUs);
